Purpose: Fix bug in SWMR object header code
Descriptions:
When removing object header messages, it is possible for object header
continuation messages to move to a different chunk.  When this happens, flush
dependencies need to be updated to reflect the new structure.  This change adds
code to update the flush dependencies, and a test for this.

Also fixed a bug where the flush dependency no the object header proxy was not
being destroyed when an object header chunk was deleted.

Tested: ummon