Purpose: Remove problematic abandonment of error stack
Description:
H5Fis_hdf5 uses H5F_locate_signature to check if the file is hdf5.  If it does
not locate the signature, H5F_locate_signature would issue an error and return
HADDR_UNDEF.  H5Fis_hdf5 does not consider it an error if the signature is not
found, so it does not issue an error or clear the stack.  The filled stack could
then cause issues later on.

Changed H5F_locate_signature to return herr_t, not issue an error if the
signature is not found, and added a parameter for a pointer to the signature
address that the function fills in.

Tested: jam, koala, ostrich (h5committest)