Port r24635 from trunk to 1.8 branch.
Note: despite the log from r24635, it was decided that this is the actual fix
for the error described in HDFFV-8647 that is not specific to BSD.  This does
not address the BSD issue, which is apparently unrelated.

Tested: jam, ostrich, platypus (h5committest, apparent system issue on koala)

Log from r24635:
Change some range check assertions to errors in H5HF_man_remove in  an attempt
to aid in diagnosing issue HDFFV-8647.

Tested: ummon